chore: stops assuming local DB
As a new postgres developer, I have been very confused as to what is happening with our local database setup and configuration. Assuming a specific db called "postgres" has made it difficult. Not to mention a database was being created, even though we are specifying (or worse, not specifying) a URL. This PR attempts to clear up what is actually happening when spinning up a local instance by expecting the developer to create their own db if they are choosing to use our DB functionality, something they should already understand, or be forced to learn about. I'll follow up with simple instructions in the template repo.